// src/Model/iTechRestApi/iTechDataTaskStatusEnum.ts
//*************************DO NOT MODIFY**************************
//
//THESE FILES ARE AUTOGENERATED WITH A TYPESCRIPT CODE GENERATION TOOL AND ANY MODIFICATIONS MADE HERE WILL BE LOST.
//PLEASE VISIT http://frhagn.github.io/Typewriter/ TO LEARN MORE ABOUT THIS VISUAL STUDIO EXTENSION
//
//*************************DO NOT MODIFY**************************



export enum iTechDataTaskStatusEnum {
    notStarted = 1,
    workingOnIt = 2,
    stuck = 3,
    waitingForReview = 4,
    ongoing = 5,
    onHold = 6,
    done = 7
}