// src/Model/iTechRestApi/iTechDataWebTabEnum.ts
//*************************DO NOT MODIFY**************************
//
//THESE FILES ARE AUTOGENERATED WITH A TYPESCRIPT CODE GENERATION TOOL AND ANY MODIFICATIONS MADE HERE WILL BE LOST.
//PLEASE VISIT http://frhagn.github.io/Typewriter/ TO LEARN MORE ABOUT THIS VISUAL STUDIO EXTENSION
//
//*************************DO NOT MODIFY**************************



export enum iTechDataWebTabEnum {
    user = 1,
    view = 2,
    template = 3,
    review = 4
}